Simplify the use-theming example
