--- how-to/use-theming/client/src/actions.ts
@@ -1,3 +1,0 @@
-export const ACTION_IDS = {
-	applyTheme: "apply-theme"
-};
